+++ tests/issue_100.spec.ts
@@ -24,7 +24,7 @@ describe('Static props prevents selectors recall (#100)', () => {
       },
     };
 
-    const selectAllBooks = memoize((state: State) => Object.values(state));
+    const selectBook1 = memoize((state: State) => state.book1);
 
     const selectPriceString = memoize(
       (state: State) => state.book1.priceString,
@@ -36,7 +36,52 @@ describe('Static props prevents selectors recall (#100)', () => {
       return priceString;
     });
 
-    selectAllBooks(state1);
+    selectBook1(state1);
+
+    expect(selectAdjustedPriceString(state1)).toBe('10');
+    expect(selectAdjustedPriceString(state2)).toBe('20');
+  });
+
+  it('case 2', () => {
+    type State = {
+      book1: {
+        staticProp: string;
+        priceString: string;
+      };
+    };
+
+    const state1: State = {
+      book1: {
+        staticProp: '5',
+        priceString: '10',
+      },
+    };
+
+    const state2: State = {
+      book1: {
+        staticProp: '5',
+        priceString: '20',
+      },
+    };
+
+    const selectBook1 = memoize((state: State) => state.book1);
+
+    const selectPriceString = memoize(
+      (state: State) => state.book1.priceString,
+    );
+
+    const selectAdjustedPriceString = memoize((state: State) => {
+      const priceString = selectPriceString(state);
+      state.book1.staticProp; // touch the prop
+      return priceString;
+    });
+
+    const selectMemoizedPriceString = memoize((state: State) =>
+      selectPriceString(state),
+    );
+
+    selectBook1(state1);
+    selectMemoizedPriceString(state1);
 
     expect(selectAdjustedPriceString(state1)).toBe('10');
     expect(selectAdjustedPriceString(state2)).toBe('20');